requirements.txt
APScheduler==3.10.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
feedparser==6.0.10
Flask==2.2.3
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pika==1.3.1
python-dotenv==1.0.0
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
sgmllib3k==1.0.0
six==1.16.0
tzdata==2023.2
tzlocal==4.3
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
